requirements.txt
pandas
openpyxl
google-generativeai
fastapi
uvicorn
pydantic
pydantic-settings